pyproject.toml
[project]
name = "pocket-musec"
version = "0.1.0"
description = "AI-powered lesson planning assistant for music teachers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pdfplumber>=0.9.0",
    "ocrmypdf>=15.0.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
pocketflow = "cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=backend --cov=cli"

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "W", "I", "N", "B", "C90"]
ignore = ["E501", "B008"]

[tool.hatch.build.targets.wheel]
packages = ["backend", "cli"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "black>=25.11.0",
    "mypy>=1.18.2",
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.4",
]
